update: triviaq questions, removed dupes update: many cleanup scripts to FluentPDO, reduce number of queries rename: column 12_hour to use_12_hour update: FluentPDO, autoprefixer, browserslist update: easier to update/add stats stats.php update: use FluentPDO bans.php, upgrade_database.php update: many timestamp columns to unsigned int add: added column to topics table change: sort order for cleanup processing cron_controller.php update: delete deleted messages also, only update caches if cache already exists Message.php update: check if torrent exists before deleting Torrent.php remove: stats table, no need to store these stats in db remove: readpost_update.php can't see any reason to reset a users read post history every 2 weeks

+++ src/Torrent.php
@@ -44,7 +44,9 @@ class Torrent
             ->where('torrent = ?', $tid)
             ->execute();
 
-        unlink(TORRENTS_DIR . $tid . '.torrent');
+        if (file_exists(TORRENTS_DIR . $tid . '.torrent')) {
+            unlink(TORRENTS_DIR . $tid . '.torrent');
+        }
         $this->clear_caches();
     }
 
@@ -261,31 +263,35 @@ class Torrent
      *
      * @return bool
      */
-    public function remove_torrent(string $infohash)
+    public function remove_torrent(string $infohash, int $tid = null, int $owner = null)
     {
         if (strlen($infohash) != 20) {
             return false;
         }
-        $torrent = $this->get_torrent_from_hash($infohash);
-        if (is_array($torrent)) {
+        if (empty($tid) || empty($owner)) {
+            $torrent = $this->get_torrent_from_hash($infohash);
+            $tid = $torrent['id'];
+            $owner = $torrent['owner'];
+        }
+        if (!empty($tid) && !empty($owner)) {
             $key = 'torrent_hash_' . bin2hex($infohash);
             $this->cache->deleteMulti([
                 $key,
-                'peers_' . $torrent['owner'],
-                'coin_points_' . $torrent['id'],
+                'peers_' . $owner,
+                'coin_points_' . $tid,
                 'latest_comments_',
                 'top5_tor_',
                 'last5_tor_',
                 'scroll_tor_',
-                'torrent_details_' . $torrent['id'],
-                'torrent_details_txt_' . $torrent['id'],
+                'torrent_details_' . $tid,
+                'torrent_details_txt_' . $tid,
                 'lastest_tor_',
                 'slider_tor_',
                 'torrent_poster_count_',
                 'torrent_banner_count_',
                 'backgrounds_',
                 'posters_',
-                'similiar_tor_' . $torrent['id'],
+                'similiar_tor_' . $tid,
             ]);
             $this->clear_caches();
         }
